test(agent): isolate Promise species mutation
